+++ qbraid_qir/profiles/__init__.py
@@ -19,7 +19,6 @@ This package exposes base and adaptive profiles, along with shared interfaces
 and registry mechanisms to manage profile capabilities and restrictions.
 """
 
-from .abstract import QIRModule, QIRVisitor
 from .core import (
     AdaptiveProfile,
     BaseProfile,
@@ -30,8 +29,6 @@ from .core import (
 )
 
 __all__ = [
-    "QIRVisitor",
-    "QIRModule",
     "ProfileCapabilities",
     "ProfileRestrictions",
     "Profile",
